Fix test DB isolation bug, stop using py.test:
Several tests that created a `django.contrib.auth.models.User` object via a
simple model viewset/serializer endpoint were working just fine.  Then as soon
as I added a test that both creates a user (either via
`django.contrib.auth.models.User.objects.create()` or via the end point) and
then *retrieves* it, all the other tests failed to create the user saying that
a user with that username already exists.  Take the create-and-retrieve test
out and the others work again.

So I put the create-and-retrieve test back in, created a simple `manage.py`
such as the one generated by `django-admin startproject`, and ran the same
tests with `python manage.py test` and they all work just fine.

So that confirms that the test runner is the culprit.  I've also run into
other issues using py.test with Django in the past.  So removing it seems
best.  This switches to using `django-setuptest`, which is test-runner
agnostic, adds standard `python setup.py test` support, and includes both
coverage and pep8 reporting.